Initial Launcher2 changes for mdpi by Revoked
Still hardcoded for a specific resolution, although that is easily
removed.  Home button is properly resized and working correctly.
Speaking with Romain Guy it seems that this application is going to
use quite a bit of video RAM, and when combined with Gallery3D or Camera
it can cause an out of memory error, causing a soft reboot.

The initial changes to make this work were completed by Revoked and have
since been tweaked and cleaned by jnwhiteh.